Move DeviceType and DeviceSubType enums to uefi-raw
And re-export them from `uefi`.
